Temporarily don't use randomness for ecdsa/es256k
`sign_digest_with_rng()` seems to be broken with the current version
of the `k256` crate.

So, fall back to deterministic signatures until a solution is found.

Do the same with `p256` (even though that one still works) for
consistency.

It's unclear wether `sign_digest_with_rng()` computes a deterministic
`k` and includes randomness, or if it directly uses the RNG output as
`k`. So, with these implementations, deterministic signatures may be
safer anyway.